apply review hints
