+++ .eslintignore
@@ -3,3 +3,4 @@
 coverage
 camunda-modeler/
 camunda-docs-manual/
+camunda-docs-static/
