+++ lib/helper/annotations.js
@@ -9,14 +9,16 @@ module.exports.addAnnotation = function addAnnotation(selector, text, options = 
   const textBox = document.createElement('div');
   textBox.style.padding = '15px 30px';
   textBox.style.backgroundColor = '#dcecff';
-  textBox.style.border = '2px solid black';
+  textBox.style.border = '1px solid black';
   textBox.style.position = 'absolute';
   textBox.style.zIndex = '9999999';
   textBox.style.fontSize = '18px';
   textBox.style.whiteSpace = 'pre-wrap';
   textBox.style.textAlign = 'center';
+  textBox.style.borderRadius = '5px';
+  textBox.style.filter = 'drop-shadow(2px 4px 6px grey)';
   textBox.classList.add('SCREENSHOT__ANNOTATION');
-  textBox.textContent = text;
+  textBox.innerHTML = text;
 
   document.body.appendChild(textBox);
 
@@ -24,6 +26,7 @@ module.exports.addAnnotation = function addAnnotation(selector, text, options = 
   // left, top: css properties of the textbox
   // c[x/y][s/e]: [s]tart/[e]nd [c]oordinates for [x]- and [y] axis of the svg arrow
   let left, top, cxs, cxe, cys, cye;
+  let markerSize = 6;
   if (Math.abs(options.x) > Math.abs(options.y)) {
     cys = bb.y + bb.height / 2;
     cye = cys + options.y;
@@ -31,12 +34,12 @@ module.exports.addAnnotation = function addAnnotation(selector, text, options = 
     if (options.x > 0) {
 
       // right
-      cxs = bb.x + bb.width + 5;
+      cxs = bb.x + bb.width + 5 + markerSize/2;
       left = cxs + options.x;
     } else {
 
       // left
-      cxs = bb.x;
+      cxs = bb.x - markerSize/2;
       left = bb.x - textBox.clientWidth + options.x - 5;
     }
     cxe = cxs + options.x;
@@ -47,12 +50,12 @@ module.exports.addAnnotation = function addAnnotation(selector, text, options = 
     if (options.y > 0) {
 
       // bottom
-      cys = bb.y + bb.height + 5;
+      cys = bb.y + bb.height + 5 + markerSize/2;
       top = cys + options.y;
     } else {
 
       // top
-      cys = bb.y - 5;
+      cys = bb.y + 2 - markerSize/2;
       top = bb.y - textBox.clientHeight + options.y;
     }
     cye = cys + options.y;
