+++ lib/quickstart.js
@@ -46,8 +46,8 @@ module.exports = async function startScreenshotBatch() {
 
       await modeler.click('[data-element-id="StartEvent_1"]');
 
-      await modeler.annotate('.djs-context-pad.open .bpmn-icon-task', 'Select activity shape (rectangle)', annotationOptions.right);
-      await modeler.annotate('[data-element-id="StartEvent_1"]', 'Double-click on element to edit label\\n(Use SHIFT + ENTER for line break)', annotationOptions.top);
+      await modeler.annotate('.djs-context-pad.open .bpmn-icon-task', 'Select <b>activity shape (rectangle)</b>', annotationOptions.right);
+      await modeler.annotate('[data-element-id="StartEvent_1"]', 'Double-click <b>Start Event</b> to edit label\\n(Use SHIFT + ENTER for line break)', annotationOptions.top);
 
       await modeler.takeScreenshot(filepath);
     }
@@ -65,7 +65,7 @@ module.exports = async function startScreenshotBatch() {
       await modeler.click('[data-action="replace"]');
       await modeler.mouseOver('[data-id="replace-with-service-task"]');
 
-      await modeler.annotate('.djs-context-pad.open', '1. Select activity\\n2. Click wrench icon and change type to Service Task', annotationOptions.right);
+      await modeler.annotate('.djs-context-pad.open', '1. Select <b>activity</b>\\n2. Click <b>wrench icon</b> and change type to <b>Service Task</b>', annotationOptions.right);
 
       await modeler.takeScreenshot(filepath);
     }
@@ -94,7 +94,7 @@ module.exports = async function startScreenshotBatch() {
       modeler = await createModeler(diagramPaths, config);
 
       await modeler.click('[data-element-id="serviceTask1"]');
-      await modeler.annotate('.properties .toggle', "If it\\'s not already visible, click \\n on the Properties Panel tab", annotationOptions.left);
+      await modeler.annotate('.properties .toggle', "If it\\'s not already visible, click \\n on the <b>Properties Panel tab</b>", annotationOptions.left);
 
       await modeler.takeScreenshot(filepath);
     }
@@ -108,9 +108,9 @@ module.exports = async function startScreenshotBatch() {
 
       modeler = await createModeler(diagramPaths, config);
 
-      await modeler.annotate('[data-entry="id"]', 'Set Process Id', annotationOptions.left);
-      await modeler.annotate('[data-entry="name"]', 'Set Process Name', annotationOptions.left);
-      await modeler.annotate('[data-entry="process-is-executable"]', 'Mark the process to be executable', annotationOptions.left);
+      await modeler.annotate('[data-entry="id"]', 'Set <b>Process ID</b>', annotationOptions.left);
+      await modeler.annotate('[data-entry="name"]', 'Set <b>Process Name</b>', annotationOptions.left);
+      await modeler.annotate('[data-entry="process-is-executable"]', 'Mark the process to be <b>executable</b>', annotationOptions.left);
 
       await modeler.takeScreenshot(filepath);
     }
@@ -202,7 +202,7 @@ module.exports = async function startScreenshotBatch() {
 
       modeler.annotate(
         '[data-action="create.task"]',
-        'Select activity shape and drag\\ninto position between Start\\nEvent and \\"Charge Credit\\nCard\\" Service Task',
+        'Select <b>activity shape</b> and drag\\ninto position between <b>Start\\nEvent</b> and <b>\\"Charge Credit\\nCard\\" Service Task</b>',
         annotationOptions.right
       );
